Add chef cookbook to provision production server
This adds a cookbook directory which is a Chef cookbook that will
hopefully be able to converge a new server to the state of our existing
one if we ever need to do so. (Run the run.sh script from the cookbook
directory to sync that directory with our server, and converge it.)

The deploy process is designed to mirror that of Capistrano, the
industry-standard way to deploy Rails applications. In Capistrano, each
server maintains a git clone, and then `git archive` is used to produce
a new copy of the code at a given ref. This copy is in a directory that
is timestamped[1]. Then, after all setup for the new version of code is
complete (in our case, `pip install`), we swap a symlink
(`/data/backend/current`) to point to the new directory.

There is a bash script, `/usr/local/bin/deploy-backend`, which does all
of the above.

Also, the cookbook installs nginx and configures it to communicate with
uwsgi. There is currently a problem passing the "Host" header to Django,
which results in all requests returning an HTTP 400 "Bad Request", but
I've gotta go and I'll have to figure that out later.

[1] So the same revision may be deployed multiple times if a rollback is
necessary.