Add mockups into source control
People need to be able to easily see what it is we're doing. Having to
search through Google Drive or Slack is not a good way to get people to
the mockups.

These will get out of date soon. But at least we will be see them as
they do.