Add django command to import from the spreadsheet
The spreadsheet is here:

  https://docs.google.com/spreadsheets/d/1272oaLyQhKwQa6RicA5tBso6wFruum-mgrNm3O3VogI/edit#gid=0

This adds a django management command which downloads the (currently,
only first page of the) spreadsheet in TSV format[1] and imports each
row as its own OfficeElection/Candidate.

  python manage.py importoaklandcandidates

There are some subsequent TODO's but for now I want to get this in so
we can start playing with real data.

Some follow-up work:
* Store the FPPC id's for candidate controlled committees and whatever
  is equivalent for ballot measures.
* Actually link up contributions with those FPPC id's
* Figure out what to do with the "Non-controlled committees"
  spreadsheet

[1]: my first attempt to parse CSV got tripped up because of commas